Add files via upload
This is a ecommerce application named YourSpace which specializes on selling hoodies and sweatshirts, It is built by using Flask, java script , html and css, the oracle dataset is used in this project for purpose of storing  and retrieving information. The entire application along with required libraries are avaiable in YourSpace_ecommerce_flask.zip file.